refactor: refactored 'Hero-Tags' logic
