+++ src/js/modules/hero-tags/index.ts
@@ -1,0 +1,3 @@
+import HeroTags from './ui/HeroTags'
+
+export default HeroTags
